requirements.txt
flask==2.3.3
elasticsearch==8.10.0
python-dotenv==1.0.0
pytest==7.4.2
requests==2.31.0
gunicorn==21.2.0
flask-cors==4.0.0 